requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
diffusers>=0.24.0
accelerate>=0.24.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0
pyyaml>=6.0.0
tqdm>=4.66.0
requests>=2.31.0


